+++ src/ViaBasicAuthorization/Resolver.php
@@ -4,12 +4,12 @@ declare(strict_types = 1);
 namespace Innmind\HttpAuthentication\ViaBasicAuthorization;
 
 use Innmind\HttpAuthentication\Identity;
-use Innmind\Immutable\Maybe;
+use Innmind\Immutable\Attempt;
 
 interface Resolver
 {
     /**
-     * @return Maybe<Identity>
+     * @return Attempt<Identity>
      */
-    public function __invoke(string $user, string $password): Maybe;
+    public function __invoke(string $user, string $password): Attempt;
 }