+++ tests/ViaFormTest.php
@@ -16,7 +16,7 @@ use Innmind\Http\{
     ProtocolVersion,
 };
 use Innmind\Url\Url;
-use Innmind\Immutable\Maybe;
+use Innmind\Immutable\Attempt;
 use PHPUnit\Framework\TestCase;
 
 class ViaFormTest extends TestCase
@@ -63,7 +63,7 @@ class ViaFormTest extends TestCase
             ->expects($this->once())
             ->method('__invoke')
             ->with($request->form())
-            ->willReturn(Maybe::just($identity = $this->createMock(Identity::class)));
+            ->willReturn(Attempt::result($identity = $this->createMock(Identity::class)));
 
         $this->assertSame($identity, $authenticate($request)->match(
             static fn($identity) => $identity,